ch4: improve tests for ch4

+++ ch4/weather/src/main/java/wf/garnier/spring/boot/test/ch4/weather/selection/CityWeather.java
@@ -3,7 +3,7 @@ package wf.garnier.spring.boot.test.ch4.weather.selection;
 import wf.garnier.spring.boot.test.ch4.weather.city.City;
 import wf.garnier.spring.boot.test.ch4.weather.openmeteo.WeatherData;
 
-public record CityWeather(String cityName, String country, Long cityId, String weather, Double temperature,
+public record CityWeather(String cityName, String country, Integer cityId, String weather, Double temperature,
 		Double windSpeed) {
 
 	public CityWeather(City city, WeatherData weatherData) {
